Move the dialogs to the main window on Linux
This solves a (the) performance problem a lot of people have been experiencing with the "basic" QSG loop model. Animations in combination with multiple QML windows eat huge amounts of CPU time. After just downloading the image, FMW was resting at about 80% CPU usage on a fairly recent/powerful processor. This was happening without it actually doing anything or animations.

It also turns out we can't force the "threaded" model because people with nouveau hardware seem to experience crashes quite consistently. There are probably other issues lying somewhere with this model too...

So until this is fixed, it seems we'll have to have the download and restore dialogs inside the main window. The user experience doesn't change much though.