Never fully fill the progress bar
so the user doesn't wonder why it shows him that it has finished writing
when it hasn't.
The bigger issue is that bytes written to the kernel cache report as
being written but actually aren't yet flushed to the disk.
There are ways to avoid this issue but it has a huge impact on the
actual writing speed so it's really not worth fixing.