Fix restore on Linux with latest UDisks/libblockdev (#647)
The newly created partition cannot start on offset 0, previously
UDisks was able to compensate for this, but there is a bug in the
latest UDisks/libblockdev and this no longer works. But in general
offset 0 for a new partition doesn't make sense. We can also use
0 for size to let UDisks calculate the maximal size instead of
passing the underlying device size.

Fixes: #644