chore: specify UAC level to `requireAdministrator` in the manifest for Windows (#700)
* helper: specify UAC level to requireAdministrator for Windows

So that the helper can trigger a UAC prompt upon launch, and the launch
attempt will fail if CreateProcess is used without admin privilege.

* src/app: specify UAC level to requireAdministrator for Windows

QProcess::start() can not start a new process with UAC level
requireAdministrator, since it is using CreateProcess. There is one API
that can trigger the UAC prompt, ShellExecute or ShellExecuteEx,
however, processes started by them can not be managed by QProcess.
There are some workarounds but since the QProcess instance is connected
to various slots, it would require a rewrite to apply the workarounds.
We simply specify the main app to require administrator privilege in
the manifest.